fix(anchor): pin the two node_anchor defaults, catch a doubled offset, wire run_main_flow, record center's memory cost
Fix round 1 on P-F Task 1 (commit 68e791f), per review-task-1.md:

1. Pin the two node_anchor defaults the plan's constraints require: the
   CLASS default stays lower_left on IoTermRef/IoTerm (fourteen existing
   construction sites supply no node sizes), run_io's own default is center
   (the flag default). Neither was pinned by any test before this commit.

2. Add a mutation-catching test for the half-cell offset: a cell whose centre
   stays inside its home region but whose lower-left + full width would not,
   so applying the offset twice (or as the whole cell instead of half) fails
   it. Verified by temporarily doubling the offset in soft_assign.anchor_xy
   and confirming this new test fails (1.0 instead of 0.0) while all 18 other
   tests in the file stayed green, exactly as review-task-1.md's I-2 measured.

3. center's ~0.87 GiB resident cost at 27.7M nodes: (a) anchor_dx/anchor_dy
   are not a duplicate of an already-held buffer -- IoTerm never stores
   node_size_x/node_size_y themselves, only the derived half-size, so there
   is nothing to dedupe against; (b) IoTerm's anchor buffers now follow
   "rects"'s existing production habit (cast to pos's dtype at each call,
   never forced to fp64) instead of anchor_offsets's fp64 default, using
   self.rects.dtype; (c) no registry code change needed -- anchor_dx/anchor_dy
   are ordinary register_buffer entries on IoTerm, and run_placement_io.py's
   existing rec.add_root("io_term", io_term) already walks them into the T2b
   lifetime scan (verified directly); (d) spike_30m_child.py now builds
   IoTerm at the --node-anchor flag default (center) with node sizes
   supplied, not the class default, so the 30M spike's resident-memory
   verdict needs re-measuring against the configuration the drivers actually
   ship (+~0.87 GiB headroom against the L4 19.5 GiB budget).

4. run_main_flow.py builds IoTerm while freeze.py already freezes on cell
   centres -- the exact three-way mismatch P-F exists to remove. Wired
   run_soft_phase/run_main_flow/build_parser/main to the same --node-anchor
   handling (default center, pin rejected before CUDA) as the other two
   drivers, matching the plan's "wired together in exactly one place per
   driver" constraint, and pinned it with a test.

+++ src/ioplace/ops/io_term.py
@@ -376,8 +376,19 @@ class IoTerm(torch.nn.Module):
         self.register_buffer("deg_bucket", torch.as_tensor(csr.deg_bucket, dtype=torch.int64,
                                                             device=device))
 
+        # P-F fix round 1 item 3(b): match "rects"'s production habit above
+        # (cast to pos's own dtype at each call, never forced to fp64) rather
+        # than anchor_offsets's fp64 default -- self.rects.dtype is the best
+        # available proxy for pos's dtype here (IoTerm never sees pos itself
+        # until forward()), and _anchor_xy/anchor_xy already re-casts to
+        # x.dtype on every use, so this only changes what's *resident*, never
+        # what's computed. An fp32 pos/rects config now gets fp32 anchor
+        # buffers instead of a forced fp64 pair (half the resident bytes);
+        # the current production configs' rects are fp64 already
+        # (rect_table's np.float64), so this is a no-op for them today.
         dx, dy = anchor_offsets(node_anchor, node_size_x, node_size_y,
-                                self.num_physical, device=device)
+                                self.num_physical, device=device,
+                                dtype=self.rects.dtype)
         self.register_buffer("anchor_dx", dx)
         self.register_buffer("anchor_dy", dy)
